chore: change client extension groupId
