+++ client/src/main/java/de/fraunhofer/iosb/ilt/client/datatransfer/DataTransferEndpoint.java
@@ -1,0 +1,71 @@
+/*
+ * Copyright (c) 2021 Fraunhofer IOSB, eine rechtlich nicht selbstaendige
+ * Einrichtung der Fraunhofer-Gesellschaft zur Foerderung der angewandten
+ * Forschung e.V.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package de.fraunhofer.iosb.ilt.client.datatransfer;
+
+import de.fraunhofer.iosb.ilt.client.ClientEndpoint;
+import jakarta.ws.rs.Consumes;
+import jakarta.ws.rs.POST;
+import jakarta.ws.rs.Path;
+import jakarta.ws.rs.PathParam;
+import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.Response;
+import org.eclipse.edc.spi.monitor.Monitor;
+
+import java.util.Objects;
+
+import static java.lang.String.format;
+
+/**
+ * Endpoint for automated data transfer
+ */
+@Consumes({MediaType.APPLICATION_JSON, MediaType.WILDCARD})
+@Produces({MediaType.APPLICATION_JSON})
+@Path(ClientEndpoint.AUTOMATED_PATH)
+public class DataTransferEndpoint {
+
+    /*
+     * Path for providers to send data to.
+     */
+    static final String RECEIVE_DATA_PATH = "receiveData";
+
+    private final Monitor monitor;
+    private final DataTransferObservable<String> observable;
+
+    DataTransferEndpoint(Monitor monitor, DataTransferObservable<String> dataTransferObservable) {
+        this.monitor = monitor;
+        this.observable = dataTransferObservable;
+    }
+
+    /**
+     * Receive data of an agreement.
+     *
+     * @param agreementId The agreement ID corresponding to the data in the request
+     *                    body.
+     * @param requestBody The asset data of a transfer request.
+     * @return OK as response.
+     */
+    @POST
+    @Path("receiveData/{agreement}")
+    public Response receiveData(@PathParam("agreement") String agreementId, String requestBody) {
+        monitor.info(format("Receiving data for agreement %s...", agreementId));
+        Objects.requireNonNull(agreementId);
+        Objects.requireNonNull(requestBody);
+        observable.update(agreementId, requestBody);
+        return Response.ok().build();
+    }
+
+}
