+++ client/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -1,1 +1,1 @@
-de.fraunhofer.iosb.client.ClientExtension
+de.fraunhofer.iosb.ilt.client.ClientExtension
